backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
spotipy>=2.23.0
torch>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.6.0
